iommu: arm-smmu: Implement NO_SMR_CHECK option
Sometimes, when we have an IOMMU configured with predefined
contexts from the bootloader, a SMR search may not find all
the usable IDs because the corresponding MASK bits may be
set, but this doesn't mean that these stream ids cannot be
used.

This is a workaround for a firmware bug, found in Qualcomm
Family-A SoCs, from MSM8994 to MSM8998, plus SDM630, SDM636
and SDM660.

Change-Id: I5aec7d5be088b36f31436603d901995434132f72