soc: qcom: msm_bus: Correctly handle probe deferral case in bus-OF
When probing msm_bus we may get a clock API -EPROBE_DEFER when
trying to get some clock for some bus node: failing the msm_bus
driver probe for this matter is stupid and counter-productive.

Return the right error number for the bus-OF error cases in the
msm_bus probe so that we can defer probing the driver if needed.
Also, while at it, clean up the msm_bus_of_to_pdata function
usage by returning an int error code instead of having to return
either a platform data struct pointer or a pointer to error code
so that the driver becomes less prone to crashing the entire
kernel for something stupid.

Change-Id: I1c540ec6a1011b6e182b8ee2b6c547e09b05c8e2