iommu: arm-smmu: Implement no ACTLR read quirk for hypervisor lockup
On some IOMMU implementations the hypervisor disallows
reading the implementation defined ACTLR registers and
trips a security fault which freezes the entire AP.

For this reason, add a DT-activated qcom,no-actlr-read quirk
to avoid doing that.
This quirk is currently known to be only needed on Qualcomm
SoCs like MSM8996, MSM8998, SDM630, SDM660 (and variants).

Change-Id: Idd97629d8426cf3cea942fbbd9fbbf8a5f695ac9