soc: qcom: msm_bus: Fix msm_bus_of_to_pdata signature for RPMH usecase
In commit 56fe5b42ad I have introduced a little restructuration
of the error handling during the probe stage, modifying the
signature of the msm_bus_of_to_pdata function.

The change went fine, but I forgot to update the function
signature in the msm_bus_rpmh header, which is used only for
the RPMh usecase.