iommu: arm-smmu: Implement QCOM MMU-500 TLB context invalidation errata
Some Qualcomm MMU-500 implementation need to execute context
invalidation in a remote spinlock: copy the implementation from
Open Devices k4.9 and fix it to use the upstream hwspinlocks
present in kernel 4.14.

Also, in the kernel 4.9 implementation this was "errata 1", but
Qualcomm has moved the min_iova align errata (errata 2 in k4.9)
as errata 1 in k4.14, so, rename the TLB errata to errata 0 in
this implementation to avoid confusion.

Change-Id: I2af9d5b5e961b809cde005f80cd48dd827a6243c