arm64: DT: MSM8998: Correctly declare irq types for remoteprocs
Remote processors are throwing out a lot of interrupts and we need
to catch them correctly from the HLOS in order to ensure a working
state without random remote subsystems crashes due to timeouts and
such.

Declare the interrupts as edge rising so that we stop relying on
bootloader (or remote) configuration of them.

Change-Id: Ief26d9c4e9747dc6639176da73a6d064b6ff7598