platform: msm: ipa_v3: Raise min freq for IPA HW v3.1
For IPA 3.1 the 37.5MHz frequency is too low and leads to a big
slow down in small data bursts and a lot higher latency in case
the interface wasn't being actively used already: in this state
the firmware responds too slow and calculates the required BW
after too much time, which is the reason of these issues.

Raise the minimum frequency for the IPAv3.1 hardware to 75MHz,
restoring a good compromise between interface responsiveness and
battery usage.