arm64: dts: msm: msm8998, cheeseburger: enable serial console output
Enable serial console both from gpio4 and from the display via the
framebuffer HACK.
By default, set the gpio as output from the serial console.

The display does not show anything, but it is a bit more
difficult to get that working at this stage. Console works as expected
from the test point (gpio4).

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>