fpc1020: add support for an enable GPIO
Some versions of the FPC1020 fingerprint reader need to be enabled using
a GPIO. OnePlus does this on a different driver ('fingerprint detect')
but it makes more sense to just do that in the fingerprint reader
driver.

On the device tree, in the node of the FP1020 reader, add the 'fpc,enable-gpio'
property to set the GPIO output high while probing the device. Example:

fpc,enable-gpio = <&tlmm 25 0x0>;

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>