fbdev: msm: introduce MDSS_BROKEN_BOOTLOADER_HANDOFF
On certain devices, the display handoff between the bootloader and the
kernel can cause issues, for example display artifacts or a long boot time.

Introduce MDSS_BROKEN_BOOTLOADER_HANDOFF to disable the handoff, forcing
the kernel to reset the display clocks and regulators instead of relying on
the setup left by the bootloader.

In this case, the continuos splash logo feature will not work and the
screen will turn itself off during boot for some seconds.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>