clk: qcom: cpu-osm-legacy: Fix speedbin parsing
Through cross-referencing with k4.4 clock-osm driver, this small
difference caused the driver in k4.14 to not register the 2.4GHz
step for the performance cluster defined through the speedbin and
cpufreq. Simply increment the k value at fmax_temp to allocate one
more data entry; this mirrors the code refactor done in the k4.4
driver.

Signed-off-by: ederekun <sedrickvince@gmail.com>